+++ ostap/math/models.py
@@ -1111,6 +1111,7 @@ for model in ( Ostap.Math.Chebyshev              ,
                Ostap.Math.BatesShape             ,
                Ostap.Math.GenPareto              , 
                Ostap.Math.ExGenPareto            , 
+               Ostap.Math.Benini                 , 
                Ostap.Math.GEV                    , 
                Ostap.Math.MPERT                  , 
                #